+++ kie-spring/src/test/resources/org/kie/spring/kie-import-releaseid.xml
@@ -17,22 +17,11 @@
 
 <beans xmlns="http://www.springframework.org/schema/beans"
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:kie="http://drools.org/schema/kie-spring" xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:kie="http://drools.org/schema/kie-spring"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-                          http://drools.org/schema/kie-spring http://drools.org/schema/kie-spring.xsd http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
+                          http://drools.org/schema/kie-spring http://drools.org/schema/kie-spring.xsd">
 
-  <kie:import releaseId-ref="namedKieSession"/>
-
-  <!--<bean id="kieSpringProcessor" class="org.kie.spring.KModuleBeanFactoryPostProcessor"/> -->
-
-  <!-- Do not modify the version: it is automatically incremented by the maven replacer plugin -->
-  <kie:releaseId id="namedKieSession" groupId="org.drools" artifactId="named-kiesession" version="6.3.0-SNAPSHOT"/>
-
-  <bean id="annotatedBean" class="org.kie.spring.beans.annotations.BeanWithReleaseId"/>
-
-  <bean id="person" class="org.kie.spring.beans.Person">
-    <constructor-arg type="java.lang.String" index="0" value="HAL"/>
-    <property name="happy" value="false"/>
-  </bean>
+  <kie:import releaseId-ref="spring-import-releaseId"/>
+  <kie:releaseId id="spring-import-releaseId" groupId="org.kie.spring" artifactId="spring-scanner-test" version="1.0-SNAPSHOT"/>
 
 </beans>
